Workflow#quick_step
Formalizes the api into the floe engine.

This hides the details of step, the constructor, and credentials from the workflow engine gem.

The goal of `quick_step` is just like step for now.

But the thought is it will call a single task and then quickly step through logic that doesn't require spawning out to tasks. So the trivial comparisons or status checks will not take multiple hits to the queue/database.