only check pod status if it is running
if it stops running, then there is no change in status, and no reason to check again.
If it is running, that has possibly changed, so check it again.

If the output was retrieved, no reason to retrieve it again.

agenda: want to externally set/override these values in error cases.
This means doing that will not undo those overrides