bin/floe parses json
The Context expects input to be a json object.
bin/floe is passing in a raw string to the Context.

BEFORE:

If there is an error in the Task, the result_path is not able to apply to the input.
This causes an error:

```
lib/floe/workflow/reference_path.rb:35:in `set': undefined method `merge!' for "{\\"foo\\": 2}":String (NoMethodError)

          result.merge!(value)
                ^^^^^^^
  from lib/floe/workflow/states/task.rb:114:in `catch_error!'
  from lib/floe/workflow/states/task.rb:55:in `finish'
  from lib/floe/workflow/state.rb:51:in `run_nonblock!'
  from lib/floe/workflow.rb:76:in `step_nonblock'
  from lib/floe/workflow.rb:63:in `step'
  from lib/floe/workflow.rb:58:in `run!'
  from exe/floe:37:in `<top (required)>'
```

After:

An error is able to be merged with the input

```
{"Error"=>"FailStateError", "Cause"=>"No Matches!"}
```