No longer store payload in States
Overview
---

We use `Payload#payload` in `State#initialize`, but then do not use it again.

This PR still passes `payload` to `initialize`, but drops the `payload` variable.

I kept `payload` around to make debugging easier, but in truth, I have never used it to debug.

Exceptions
---

We do need `Workflow#payload` to validate the state references in `"Next"`.
If we store/pass `Workflow#payload["State"].keys` into child initializers, then this can be dropped.

We do need `Path#payload`, but this is not a payload but rather a path location.